% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{exampleAnalysisData}
\alias{exampleAnalysisData}
\title{Load the Synthetic Example Analysis Data}
\usage{
exampleAnalysisData()
}
\value{
A \code{MultiAssayExperiment}.
}
\description{
Load the Synthetic Example Analysis Data
}
\examples{
analysisData <- exampleAnalysisData()
names(MultiAssayExperiment::experiments(analysisData))
}
