% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{exampleKnowledgeNetwork}
\alias{exampleKnowledgeNetwork}
\title{Load the Synthetic Example Knowledge Network}
\usage{
exampleKnowledgeNetwork()
}
\value{
A standardized knowledge-network \code{DataFrame}.
}
\description{
Load the Synthetic Example Knowledge Network
}
\examples{
knowledgeNetwork <- exampleKnowledgeNetwork()
nrow(knowledgeNetwork)
}
