% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{exampleSeedNodes}
\alias{exampleSeedNodes}
\title{Load the Synthetic Example Seed Nodes}
\usage{
exampleSeedNodes()
}
\value{
A character vector of seed-node identifiers.
}
\description{
Load the Synthetic Example Seed Nodes
}
\examples{
seedNodes <- exampleSeedNodes()
seedNodes
}
