% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-data.R
\name{validateAnalysisData}
\alias{validateAnalysisData}
\title{Validate a CorNetto Analysis Object}
\usage{
validateAnalysisData(analysisData, quiet = FALSE)
}
\arguments{
\item{analysisData}{A \code{MultiAssayExperiment}.}

\item{quiet}{Whether to suppress the success message.}
}
\value{
The validated \code{MultiAssayExperiment}.
}
\description{
Validate that a \code{MultiAssayExperiment} is compatible with CorNetto.
}
\examples{
analysisData <- exampleAnalysisData()
validated <- validateAnalysisData(analysisData)
validated
}
